Add CBA settings file, with some settings
This file will be read and its settings applied in the mission. The reason for adding it is that the addon options UI is difficult to navigate to see what settings have been edited, compared to reading the file.

Additionally, the disable remote sensors setting has been enabled, which may lead to performance gains at the cost of increased vulnerability for locality issues when using scripts that rely on knowsAbout.

ACE Weather settings have also been included, to make it more clear that it more easily toggleable for missions.